refactor: Mapper 위치 변경
